pyproject.toml
[tool.poetry]
name = "profanity-filter-docker"
version = "0.1.0"
description = ""
authors = ["huscker <egostalker03@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
profanity-filter = {git = "https://github.com/huscker/profanity-filter.git", extras = ["deep-analysis", "pymorphy2-ru"], rev="7a23c3204d482196c6a60b364c21690f6a5b3afe"}
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
fastapi-admin = "^1.0.4"
tortoise-orm = "^0.20.0"
prometheus-fastapi-instrumentator = "^7.0.0"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.5.post1"
pytest = "^8.1.1"
aerich = "^0.7.2"
lazy-object-proxy = "^1.10.0"
asyncpg = "^0.29.0"
aiocache = "^0.12.2"
factory-boy = "^3.3.0"
factory-boy-extra = "^0.1.3"
numpy = "<2.0.0"

[tool.aerich]
tortoise_orm = "service.config.orm_config"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
